rag-Pinecone/requirements.txt
langchain
langchain-groq
groq
numpy
langchain-community 
langchain-huggingface
langchain-postgres
langchain-text-splitters
sentence-transformers
psycopg-binary
langchain-pinecone
pinecone-client